Fixes bug with CashIssueFlow in Java API.
