+++ java-source/build.gradle
@@ -7,7 +7,7 @@ repositories {
     maven { url 'https://ci-artifactory.corda.r3cev.com/artifactory/corda-releases' }
 }
 
-apply plugin: 'java'
+apply plugin: 'kotlin'
 apply plugin: 'net.corda.plugins.cordformation'
 apply plugin: 'net.corda.plugins.quasar-utils'
 
@@ -23,10 +23,10 @@ sourceSets {
         }
     }
     integrationTest {
-        java {
+        kotlin {
             compileClasspath += main.output + test.output
             runtimeClasspath += main.output + test.output
-            srcDir file('src/integrationTest/java')
+            srcDir file('src/integrationTest/kotlin')
         }
     }
 }
@@ -37,6 +37,8 @@ configurations {
 }
 
 dependencies {
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
+    testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
     testCompile "junit:junit:$junit_version"
 
     // Corda integration dependencies
@@ -58,8 +60,13 @@ dependencies {
     cordapp "net.corda:corda-finance:$corda_release_version"
 }
 
-tasks.withType(JavaCompile) {
-    options.compilerArgs << "-parameters" // Required for shell commands.
+tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
+    kotlinOptions {
+        languageVersion = "1.1"
+        apiVersion = "1.1"
+        jvmTarget = "1.8"
+        javaParameters = true   // Useful for reflection.
+    }
 }
 
 task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
