--- java-source/src/main/java/net/corda/examples/obligation/ObligationBaseFlow.java
@@ -1,61 +1,0 @@
-package net.corda.examples.obligation;
-
-import com.google.common.collect.ImmutableList;
-import net.corda.core.contracts.StateAndRef;
-import net.corda.core.contracts.UniqueIdentifier;
-import net.corda.core.flows.FlowException;
-import net.corda.core.flows.FlowLogic;
-import net.corda.core.flows.FlowSession;
-import net.corda.core.flows.SignTransactionFlow;
-import net.corda.core.identity.AbstractParty;
-import net.corda.core.identity.Party;
-import net.corda.core.node.services.Vault;
-import net.corda.core.node.services.vault.QueryCriteria;
-import net.corda.core.transactions.SignedTransaction;
-import net.corda.core.utilities.ProgressTracker;
-
-import java.util.List;
-
-/**
- * An abstract FlowLogic class that is subclassed by the obligation flows to
- * provide helper methods and classes.
- */
-abstract class ObligationBaseFlow extends FlowLogic<SignedTransaction> {
-
-    Party getFirstNotary() throws FlowException {
-        List<Party> notaries = getServiceHub().getNetworkMapCache().getNotaryIdentities();
-        if (notaries.isEmpty()) {
-            throw new FlowException("No available notary.");
-        }
-        return notaries.get(0);
-    }
-
-    StateAndRef<Obligation> getObligationByLinearId(UniqueIdentifier linearId) throws FlowException {
-        QueryCriteria queryCriteria = new QueryCriteria.LinearStateQueryCriteria(
-                null,
-                ImmutableList.of(linearId),
-                Vault.StateStatus.UNCONSUMED,
-                null);
-
-        List<StateAndRef<Obligation>> obligations = getServiceHub().getVaultService().queryBy(Obligation.class, queryCriteria).getStates();
-        if (obligations.size() != 1) {
-            throw new FlowException(String.format("Obligation with id %s not found.", linearId));
-        }
-        return obligations.get(0);
-    }
-
-    Party resolveIdentity(AbstractParty abstractParty) {
-        return getServiceHub().getIdentityService().requireWellKnownPartyFromAnonymous(abstractParty);
-    }
-
-    static class SignTxFlowNoChecking extends SignTransactionFlow {
-        SignTxFlowNoChecking(FlowSession otherFlow, ProgressTracker progressTracker) {
-            super(otherFlow, progressTracker);
-        }
-
-        @Override
-        protected void checkTransaction(SignedTransaction tx) {
-            // TODO: Add checking here.
-        }
-    }
-}
