+++ kotlin-source/src/main/kotlin/net/corda/examples/obligation/ObligationApi.kt
@@ -5,6 +5,9 @@ import net.corda.core.contracts.UniqueIdentifier
 import net.corda.core.messaging.CordaRPCOps
 import net.corda.core.utilities.OpaqueBytes
 import net.corda.core.utilities.getOrThrow
+import net.corda.examples.obligation.flows.IssueObligation
+import net.corda.examples.obligation.flows.SettleObligation
+import net.corda.examples.obligation.flows.TransferObligation
 import net.corda.finance.contracts.asset.Cash
 import net.corda.finance.contracts.getCashBalances
 import net.corda.finance.flows.CashIssueFlow
