+++ kotlin-source/src/test/kotlin/net/corda/examples/obligation/NodeDriver.kt
@@ -10,7 +10,7 @@ import net.corda.testing.driver.driver
 fun main(args: Array<String>) {
     val user = User("user1", "test", permissions = setOf())
 
-    driver(isDebug = true) {
+    driver(isDebug = true, startNodesInProcess = true) {
         startNode(
                 providedName = CordaX500Name("Controller", "London", "GB"),
                 advertisedServices = setOf(ServiceInfo(SimpleNotaryService.type))
